Allow null values in FileField@getDatabaseValue. Fixes #78

+++ src/app/Library/Fields/FileField.php
@@ -44,6 +44,10 @@ class FileField extends InputField
      */
     public function getDatabaseValue($value): string
     {
+        if (! $value instanceof UploadedFile) {
+            return parent::getDatabaseValue($value);
+        }
+
         $path = $this->saveFile($value);
 
         $this->deleteFile();
